requirements.txt
streamlit
streamlit_authenticator
pandas
numpy
plotly
scikit_learn
fpdf2
openpyxl
kaleido==0.2.1
pillow
pdfplumber
